requirements.txt
Pillow
pyqt5
moviepy
numpy